Commit to pull

+++ django-project/courses/additional_pages/actions.py
@@ -20,7 +20,8 @@ def add(request):
     staging_page = AdditionalPage(course=common_page_data['staging_course'], title=request.POST.get("title"), slug=request.POST.get("slug"), index=index, mode='staging')
     staging_page.save()
     
-    staging_page.create_production_instance()
+    # @deprecated: We want production instance to be created on first publish
+    #staging_page.create_production_instance()
     
     return redirect(request.META['HTTP_REFERER'])
     
@@ -69,6 +70,7 @@ def delete(request):
         return
         
     page.delete()
-    page.image.delete()
+    if page.image:
+        page.image.delete()
     
     return redirect('courses.additional_pages.views.manage_nav_menu', common_page_data['course_prefix'],common_page_data['course_suffix'])
