+++ django-project/courses/announcements/actions.py
@@ -24,7 +24,9 @@ def add_announcement(request):
         owner=request.user,
     )
     announcement.save()
-    announcement.create_production_instance()
+    
+    # @deprecated: We want production instance to be created on first publish
+    #announcement.create_production_instance()
     
     if request.POST.get("commit") == '1':
         announcement.commit()
@@ -64,7 +66,9 @@ def delete_announcement(request):
 
     announcement = Announcement.objects.get(id=request.POST.get("announcement_id"))
     announcement.delete()
-    announcement.image.delete()
+    
+    if announcement.image:
+        announcement.image.delete()
     
     return redirect('courses.announcements.views.list', request.POST.get("course_prefix"), request.POST.get("course_suffix"))
